+++ src/main/java/module-info.java
@@ -1,0 +1,20 @@
+module data.api {
+	exports io.github.astrapi69.data.accomplishable;
+	exports io.github.astrapi69.data.activatable;
+	exports io.github.astrapi69.data.auditable;
+	exports io.github.astrapi69.data.creatable;
+	exports io.github.astrapi69.data.deletable;
+	exports io.github.astrapi69.data.enumtype;
+	exports io.github.astrapi69.data.identifiable;
+	exports io.github.astrapi69.data.modifiable;
+	exports io.github.astrapi69.data.nameable;
+	exports io.github.astrapi69.data.pairable;
+	exports io.github.astrapi69.data.processable;
+	exports io.github.astrapi69.data.traceable;
+	exports io.github.astrapi69.data.treeable;
+	exports io.github.astrapi69.data.validatable;
+	exports io.github.astrapi69.data.validatable.versionable;
+	exports io.github.astrapi69.data.verifiable;
+	exports io.github.astrapi69.data.versionable;
+	exports io.github.astrapi69.data.visibility;
+}